requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
python-dateutil==2.8.2
pydantic>=2.0.0
loguru>=0.7.0
redis>=4.5.0
nltk>=3.8.1
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0